Se agrega avance proyecto

+++ src/main/java/cl/proyecto/clientes/controller/ClienteController.java
@@ -2,26 +2,38 @@ package cl.proyecto.clientes.controller;
 
 import cl.proyecto.clientes.model.entity.Cliente;
 import cl.proyecto.clientes.service.IClienteService;
+import cl.proyecto.clientes.service.IUploadFileService;
+import lombok.extern.slf4j.Slf4j;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.Resource;
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
 
 import javax.validation.Valid;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @CrossOrigin(origins = {"http://localhost:4200", "http://localhost:65163"})
 @RestController
+@Slf4j
 @RequestMapping("/apiCliente")
 public class ClienteController {
+    private static final Logger logger = LoggerFactory.getLogger(ClienteController.class);
+
+    @Autowired
+    private IUploadFileService fotoService;
 
     @Autowired
     private IClienteService iClienteService;
@@ -129,6 +141,11 @@ public class ClienteController {
     public ResponseEntity<?> delete(@PathVariable Long id){
         Map<String, Object> respuesta = new HashMap<>();
         try {
+            Cliente clienteActual = iClienteService.findById(id);
+            if (!fotoService.eliminarArchivo(clienteActual.getFoto())) {
+                respuesta.put("mensaje", "Error al eliminar imagen antigua: ".concat(clienteActual.getFoto()));
+                return new ResponseEntity<Map<String, Object>>(respuesta, HttpStatus.INTERNAL_SERVER_ERROR);
+            }
             iClienteService.delete(id);
         }catch (DataAccessException e){
             respuesta.put("mensaje" , "Error al eliminar cliente: ".concat(e.getMessage()));
@@ -138,4 +155,53 @@ public class ClienteController {
         respuesta.put("mensaje" , "Cliente eliminado con éxito!");
         return new ResponseEntity<Map<String, Object>>(respuesta, HttpStatus.OK);
     }
+
+    @PostMapping("/clientes/upload")
+    public ResponseEntity<?> upload(@RequestParam("imagenUsuario") MultipartFile imagen, @RequestParam("id") Long id){
+        Map<String, Object> respuesta = new HashMap<>();
+        Cliente clienteActual = iClienteService.findById(id);
+
+        if (!imagen.isEmpty()){
+            String nombreArchivo = null;
+            try {
+                nombreArchivo = fotoService.copiarArchivo(imagen);
+            } catch (IOException e) {
+                respuesta.put("mensaje" , "Error subir imagen: ".concat(imagen.getOriginalFilename()));
+                respuesta.put("error" , "Descripcion error: ".concat(e.getCause().getMessage()));
+                return new ResponseEntity<Map<String, Object>>(respuesta, HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+            if (!fotoService.eliminarArchivo(clienteActual.getFoto())) {
+                respuesta.put("mensaje", "Error al eliminar imagen antigua: ".concat(clienteActual.getFoto()));
+                return new ResponseEntity<Map<String, Object>>(respuesta, HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+
+            clienteActual.setFoto(nombreArchivo);
+            iClienteService.save(clienteActual);
+
+            respuesta.put("cliente" , clienteActual);
+            respuesta.put("mensaje", "Imagen " + nombreArchivo + " subida correctamente.");
+
+        }
+
+        return new ResponseEntity<Map<String, Object>>(respuesta, HttpStatus.CREATED);
+    }
+
+    @GetMapping("/uploads/img/{nombreFoto:.+}")//expresion regular que indica que parametro trae una extension
+    public ResponseEntity<Resource> verFoto(@PathVariable String nombreFoto){
+
+        Resource recurso = null;
+        try {
+            recurso = fotoService.cargarArchivo(nombreFoto);
+        } catch (MalformedURLException e) {
+            logger.error("Error al cargar archivo " + e.getMessage());
+        }
+
+        HttpHeaders cabecera = new HttpHeaders();
+        //Fuerza que la imagen se descargue
+        cabecera.add(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + recurso.getFilename() + "\"");
+
+        return new ResponseEntity<Resource>(recurso, cabecera, HttpStatus.OK);
+    }
+
+
 }
